Move Source/stb_image.h to Dependencies directory
