+++ CMakeLists.txt
@@ -38,6 +38,7 @@ add_library(OsirisOBJ OBJECT ${SOURCES})
 set_target_properties(OsirisOBJ PROPERTIES POSITION_INDEPENDENT_CODE ON)
 target_precompile_headers(OsirisOBJ PRIVATE "${CMAKE_SOURCE_DIR}/Source/pch.h")
 target_include_directories(OsirisOBJ PUBLIC "${CMAKE_SOURCE_DIR}/Source")
+target_include_directories(OsirisOBJ PUBLIC "${CMAKE_SOURCE_DIR}/Dependencies")
 
 if (NOT WIN32)
   find_package(Freetype REQUIRED)
