+++ Source/InventoryChanger/InventoryChanger.cpp
@@ -11,7 +11,7 @@
 #define STBI_NO_FAILURE_STRINGS
 #define STBI_NO_STDIO
 #define STB_IMAGE_IMPLEMENTATION
-#include "../stb_image.h"
+#include <stb_image.h>
 
 #include "../imgui/imgui.h"
 #define IMGUI_DEFINE_MATH_OPERATORS
