+++ Source/Osiris.vcxproj.filters
@@ -550,9 +550,6 @@
     <ClInclude Include="SDK\ViewRenderBeams.h">
       <Filter>SDK</Filter>
     </ClInclude>
-    <ClInclude Include="stb_image.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="SDK\FileSystem.h">
       <Filter>SDK</Filter>
     </ClInclude>
